fix propTypes definition in List

+++ src/components/List/List.js
@@ -7,7 +7,7 @@ List.propTypes = {
   className: PropTypes.string,
   items: PropTypes.arrayOf(PropTypes.exact({
     id: PropTypes.string.isRequired,
-    label: PropTypes.string.isRequired,
+    text: PropTypes.string.isRequired,
   })).isRequired,
   onDeleteButtonClick: PropTypes.func.isRequired,
 };
@@ -22,9 +22,11 @@ function List({className, items, onDeleteButtonClick, ...props}) {
       ].join(' ')}
     >
       {
-        items.map(item => {
+        items?.map(item => {
 
           function handleOnClick(event) {
+            if (!onDeleteButtonClick) return;
+
             onDeleteButtonClick(event, item.id);
           }
